// apps/files/src/dto/entity-property-dto.ts
export interface EntityPropertyDto {
  category: String;
  serviceTcpHost: String;
  serviceTcpPort: String;
  // entityGroup: String;
  // entityName: String;
  // maxFileSizeBytes: Number;
  // acceptedType: String[];
  // volatileAtInitialized: Boolean;
}

export interface EntityPropertySearchDto {
  category: String;
}